Separate custom ops tests from caffe2
